tests/wpt/web-platform-tests/css/css-box/parsing/min-width-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing min-width with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-min-width">
<meta name="assert" content="min-width supports the full grammar '[ [<length> | <percentage>] && [border-box | content-box]? ] | available | min-content | max-content | fit-content'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("min-width", "10px");
test_valid_value("min-width", "20%");
test_valid_value("min-width", "calc(2em + 3ex)");

test_valid_value("min-width", "min-content");
test_valid_value("min-width", "max-content");
test_valid_value("min-width", "fit-content");

// The following are not yet supported by browsers:
// test_valid_value("min-width", "available");
// test_valid_value("min-width", "10px border-box");
// test_valid_value("min-width", "content-box 20%", "20% content-box");
</script>
</body>
</html>
